pp_ctl.c: Remove PL_curcop assignment
This was added years ago in commit 7fb6a879.  Now that save_re_context
does nothing (and die_unwind no longer even calls it), we don’t need to
assign to PL_curcop.  In the case where it matters, that is, the implicit
eval scope that require provides, the values of PL_curcop and oldcop
should always be the same.