Semicolon before ellipsis inside block disambiguates.
Correct documentation which indicated that, inside a block, a semicolon after
an ellipsis statement would disambiguate between a block and a hash reference
constructor.  The semicolon must precede the ellipsis to perform this
disambiguation.

Add tests to demonstrate that whitespace around the ellipsis statement does
not impeded the disambiguation.  Add perldelta entry.

For: RT #122661