Disable isfinite on VMS.
While the macro exists (in fp.h, not math.h) and it does work on
ordinary doubles, it does not work on long doubles, which is the
only use of it so far in the Perl core.  In Perl_sv_vcatpvfn_flags,
isfinite is testing a long double even when not configured with
long doubles (assuming long doubles are at least available).

So the better part of valor is to say we don't have it and use the
fallback.  Gets t/op/infnan.t #13 passing.