Change IS_PADGV to isGV in op.c:newPADOP
IS_PADGV is about to go bye-bye.  IS_PADGV was returning true only
for GVs marked as being present in the pad.  But since any GV can
make its way in there (just mention *foo under threads), it’s actually
somewhat meaningless.

This check in newPADOP is perhaps useless, too, since it is only called
by newGVOP in core (so sv is always a GV), and by B::Generate on CPAN,
which, again, gives it a GV.