pad.c: Stop turning on GvIN_PAD
This flag never really made any sense.  It allows us to identify
whether a GV is being used by a pad (e.g., for *foo under ithreads),
but *any* GV could have this set and there is no way to know whether
it was set for the pad we are currently looking at.