Only #define IS_(PADGV|CONST) if !PERL_CORE
and give IS_PADGV a simpler definition.

These are not used in the perl core any more and shouldn’t be.

The IS_PADGV definition checked for the IN_PAD flag, which flag never
made much sense (see the prev. commit’s message).  Since any GV could
end up with that flag, and since any GV coming near a pad would get
it, it might as well have been turned on for all GVs (except copies).
So just check whether the thingy is a GV.