toke.c: Fix bad read in S_scan_heredoc
If the first character of a source line matched the first character
of the here-doc terminator, then we did a memEQ with the length of
the terminator, without regard to the length of the line, which
could be shorter.

This bug goes all the way back to perl 3.000, which introduced here-
docs.  Line 1802 of toke.c in perl 3.000 is:

		    if (*s == term && bcmp(s,tokenbuf,len) == 0) {

and contains the same bug.

So this bug is twenty-five years old.