Disable libquadmath in configure.com.
It can't reasonably be a configuration option since there is no
special library needed (or available) to do 128-bit doubles.

Also, we can't set the default of a configure variable to "n" unless
it's the default answer to a question we ask; the compiler
understands "#define" and "#undef" but not "#n".