regcomp.c: Use strnEQ instead of series of tests
At some point it is more efficient and certainly clearer to call a
library function to do a strcmp than to test each individual byte.
Based on irc discussion with the original coder of this, I changed this
sequence of 6 comparisions to a strnEQ.  At the same time, this adds an
buffer overlflow check.  I was unable to cause the previous code to
overflow, but I believe it was possible.  And this changes the magic
number 6 in the code to a mnemonic giving its meaning.