Suppress some Solaris warnings
We get an integer overflow message when we left shift a 1 into the
highest bit of a word.  This changes the 1's into 1U's to indicate
unsigned.  This is done for all the flag bits in the affected word, as
they could get reorderd by someone in the future, unintentionally
reintroducing this problem again.