regcomp.c: fix fencepost error in vFAIL at end of pattern
If RExC_parse was after RExC_end we passed a negative length
into the UTF8f macro/format, which then triggered asserts.

As part of this we add a new diagnostic to show when a
(?(condition)...) construct has gone unterminated, and tests for
the unterminated case, thus testing that vFAIL works properly
at the end of the string at the same time.