Stop setting PADMY; renumber PADSTALE
The PADMY flag was originally used on values stored in pads as a way
to mark those slots ase being in use already during pad allocation.
That changed for the most part all the way back in bbce6d6978
(perl5.003_09), but vestiges still remained, because some ops used
PADMY for their targets.  I removed the last one yesterday in
14d91147.  So the PADMY flag now serves no purpose.

At run time, the sole purpose of PADMY is to determine the meaning of
the flag bit shared by PADTMP and PADSTALE.  If PADMY is set, the flag
means the latter.  Instead of that more complicated check, we can just
renumber PADSTALE to use the PADMY bit and assume that anything not
PADTMP is PADMY.

This commit changes the flags and does just enough to get
tests passing (except Peek.t).

fixup for padmy flag renumbering