Add flags to cv_name; allow unqualified retval
One of the main purposes of cv_name was to provide a way for CPAN mod-
ules easily to obtain the name of a sub.  As written, it was not
actually sufficient, as some modules, such as Devel::Declare, need an
unqualified name.

So I am breaking compatibility with 5.21.4 (which introduced cv_name,
but is only a dev release) by adding a flags parameter.