Fix assertion failure with ... ? \$a : \$b = ...
Because both branches have a srefgen, this was treated as a scalar
reference assignment, which is usually handled by refassign.  But
refassign expects the thing on the lhs to be srefgen, not a cond_expr.

The logic for propagating the assignment type must not propagate the
‘ASSIGN_REF’ type outward.  Only list context propagates outward.
Eventually, this will be handled by sassign and the cond_expr will
have lvref kids.  (It’s too complicated for refassign, which is the
optimised form that combines \ and = into one op.)

For now, this commit just fixes the assertion failure without making
this type of assignment work yet.