Some NV_TO_FV fixes following fdb7e3a630fea81.
I had missed the case where we pull an argument off the stack with
va_arg, and in order to avoid calling va_arg twice we now have to
save the double value in a local variable.  Moved SvNV out of this
macro so it's dealing exclusively with C types.

t/op/infnan.t is now passing all tests on VMS for the ordinary
double case (not -Duselongdouble).