regcomp.c: Properly dereference a ptr
UTF8_IS_INVARIANT takes a byte as its argument, not a ptr.  This bug was
introduced by 62a59291 in 5.21.4.  It doesn't probably cause bugs
because the ptr is always too big to be an invariant, but it slows
things down somewhat by not taking a shortcut it could.

It turns out that there is a discrepency between the APIs of the
UTF8-foo macros.  Some take a byte, and some a string ptr.