Fix start-of-string error in tovmsspec.
When translating a filespec starting with an extended character,
say C<+foo>, from Unix to VMS format, we were inadvertently
skipping over the first character and thus not translating it to
C<^+foo> with the caret escape prepended to the plus sign as we
should have been.

This fixes a fairly new test failure in ExtUtils::Manifest where
a filename starts with a single quote.