Deparse \(..., (@a), ...) correctly
It is the presence of parentheses immediately around the array (and
possibly local/our/my/+) in refgen context, rather than parentheses
surrounding the argument to \, that causes the array to be flattened.

B::Deparse was deparsing \(@a) correctly, but not \(@a, (@b)), in
which takes a reference to @a and to each of @b’s elements.  It was
omitting the parentheses around @b.