Fix assertion failure/hang with / (?{(^{})/
When this invalid construct is parsed, the resulting op tree for the
pattern has a code block with no constant item following it, breaking
the assumptions made by pmruntime.

Fixing this was not so easy.

You can’t just adjust the assertions, because the hang that non-debug-
ging builds exhibited is still there.

You can’t just return NULL from pmruntime when encounting the bad op
tree, because the parser will crash on the null pointer.

You can’t just return the empty pmop, because the wrong pad is
active, and other functions in op.c will try to access nonexistent
pad entries.

You can’t just LEAVE_SCOPE and return the pmop, because then PL_parser
will be null in yyerror.  Changing yyerror to account is not suffi-
cient, because then you get double-freed SVs.  At that point I gave up
with that approach.

The easiest solution turned out to be to fake up the op that we were
expecting to see.