Simplify OP_LEAVESUBLV logic in op.c:op_lvalue
If the lvalue type was OP_LEAVESUBLV, then the code for handling
entersub ops would go through extra checking to see whether we
have an lvalue sub available at compile time, before falling
through to the croaking code.  But the croaking code does not croak
if the type is OP_LEAVESUBLV, so the extra checks were for nothing.
We can skip those checks here just as we do for OP_GREPSTART,
OP_ENTERSUB and OP_REFGEN lvalue types.