Deparse sub calls quietly
In commit a958cfbb8 I used eval{prototype "CORE::..."} to see whether
something is a keyword, not realising that we have a __DIE__ hand-
ler that squeaks every time.  It is only installed when Deparse is
invoked via -MO.

Localising $SIG{__DIE__} does the trick.  Localising $@ as well is a
good idea when using eval, though I haven’t checked whether that was
causing problems.