\N{...} now treated as literal for EBCDIC qr/[...]/
Perl specially handles some ranges in bracketed character classes to
deal with gaps between, for example, "i" and "j".  The next commit will
better document this.  This special handling is only invoked when both
ends of the range are literals.  This commit extends that so that a
\N{...} construct is considered a literal for this purpose.

There really shouldn't be any difference in the handling of "A" vs
"\N{LATIN CAPITAL LETTER A}".  And this commit makes them be handled the
same.

It further makes \N{U+0041} be treated the same.  The programmer here is
effectively saying "A", and invoking the Unicode definition for it, so
it stands to reason that the range s/he wants should also be interpreted
the Unicode way.