op.c:newASSIGNOP: Rmv LINKLIST call
Originally, aassign_common_vars used LINKLIST itself because it
was following op_next pointers.  That changed in 3023b5f30, so the
LINKLIST was moved outside that function.

The LINKLIST call could not simply have been deleted, because the
split optimisation that follows assumes that it has been called and
fixes up an op_next pointer.

If we remove the LINKLIST call *and* the op_next fix-up, then every-
thing works.  The op_next fix-up was only happening to begin with
because of LINKLIST.