Set PERL_MM_USE_DEFAULT in make_ext.pl
In the perl core, we shouldn’t be getting messages like this when
trying to build perl:

Should I do external tests?
These tests will fail if there is no internet connection or if a firewall
blocks or modifies some traffic.
[y/N] [n]

libnet’s Makefile.PL now does that.  To make it easier to maintain
such Makefiles.PL to work both on CPAN and in the core, set
PERL_MM_USE_DEFAULT temporarily in make_ext.pl when running the
Makefile.PL.